Backend/requirements.txt
Django==5.0.6

# Optional but recommended for running the model locally
transformers==4.41.2
accelerate==0.30.1
bitsandbytes==0.43.1
peft==0.11.1
datasets==2.19.0
huggingface_hub==0.24.3
Pillow==10.4.0
requests==2.32.3
